Do not generate unused records in specs for DataMapper
It takes a lot of time executing specs for DataMapper. This is because we generate 100 user records and 50 project records for the first user. However, that projects data is actually not used in most of the specs. So do not create them if unneeded and do it only for specs that really need them.